skills/gitlab-ci-skill/requirements.txt
# GitLab CI Skill 依赖
python-gitlab>=4.0.0
PyYAML>=6.0

# 开发依赖
pytest>=7.0.0
pytest-cov>=4.0.0
